+++ interface.go
@@ -3,9 +3,9 @@ package annindex
 import "encoding/gob"
 
 type ANNIndex interface {
-	Train(data []float64) error
-	Add(data []float64) error
-	Search(query []float64, k int) ([][]int, error)
+	Train(data []float32) error
+	Add(data []float32) error
+	Search(query []float32, k int) ([][]int, error)
 	NumVectors() int
 	Encode(enc *gob.Encoder) error
 	Decode(dec *gob.Decoder) error
